Switch between apiClients based on msalEnabled
