+++ Src/WitsmlExplorer.Frontend/services/apiClient.tsx
@@ -1,116 +1,39 @@
-import fetch from "isomorphic-unfetch";
-import CredentialsService, { BasicServerCredentials } from "./credentialsService";
+import { msalEnabled } from "../msal/MsalAuthProvider";
+import BasicClient from "./basicClient";
+import { BasicServerCredentials } from "./credentialsService";
+import MsalApiClient from "./msalApiClient";
 
-export default class ApiClient {
-  static getCommonHeaders(credentials: BasicServerCredentials[]): HeadersInit {
-    const hasCredentials = credentials[0] !== undefined && credentials[0].password !== undefined;
-    return {
-      "Content-Type": "application/json",
-      ...(hasCredentials ? { Authorization: this.createAuthorizationString(credentials) } : {}),
-      "Witsml-ServerUrl": credentials[0]?.server?.url.toString() ?? "",
-      "Witsml-Source-ServerUrl": credentials[1]?.server?.url.toString() ?? ""
-    };
-  }
+export let ApiClient: IApiClient;
 
-  private static createAuthorizationString(credentials: BasicServerCredentials[]): string {
-    const credentialsStrings = credentials.map(({ username, password }) => `${username}:${password}`);
-    return "Basic " + btoa(credentialsStrings.join(":"));
-  }
+if (msalEnabled) {
+  ApiClient = MsalApiClient;
+} else {
+  ApiClient = BasicClient;
+}
 
-  public static async get(
-    pathName: string,
-    abortSignal: AbortSignal | null = null,
-    authConfig = AuthConfig.WITSML_AUTHENTICATION_REQUIRED,
-    currentCredentials = CredentialsService.getCredentials()
-  ): Promise<Response> {
-    const requestInit: RequestInit = {
-      signal: abortSignal,
-      headers: ApiClient.getCommonHeaders(currentCredentials)
-    };
+export interface IApiClient {
+  get: (pathName: string, abortSignal?: AbortSignal | null, currentCredentials?: BasicServerCredentials[], authConfig?: AuthConfig) => Promise<Response>;
+  post: (pathName: string, body: string, abortSignal?: AbortSignal | null, currentCredentials?: BasicServerCredentials[], authConfig?: AuthConfig) => Promise<Response>;
+  patch: (pathName: string, body: string, abortSignal?: AbortSignal | null, authConfig?: AuthConfig) => Promise<Response>;
+  delete: (pathName: string, abortSignal?: AbortSignal | null, authConfig?: AuthConfig) => Promise<Response>;
+}
 
-    return ApiClient.runHttpRequest(pathName, requestInit, authConfig);
-  }
-
-  public static async post(
-    pathName: string,
-    body: string,
-    abortSignal: AbortSignal | null = null,
-    authConfig = AuthConfig.WITSML_AUTHENTICATION_REQUIRED,
-    currentCredentials: BasicServerCredentials[] = CredentialsService.getCredentials()
-  ): Promise<Response> {
-    const requestInit = {
-      signal: abortSignal,
-      method: "POST",
-      body: body,
-      headers: ApiClient.getCommonHeaders(currentCredentials)
-    };
-    return ApiClient.runHttpRequest(pathName, requestInit, authConfig);
-  }
-
-  public static async patch(pathName: string, body: string, abortSignal: AbortSignal | null = null, authConfig = AuthConfig.WITSML_AUTHENTICATION_REQUIRED): Promise<Response> {
-    const currentCredentials = CredentialsService.getCredentials();
-    const requestInit = {
-      signal: abortSignal,
-      method: "PATCH",
-      body: body,
-      headers: ApiClient.getCommonHeaders(currentCredentials)
-    };
-
-    return ApiClient.runHttpRequest(pathName, requestInit, authConfig);
-  }
-
-  public static async delete(pathName: string, abortSignal: AbortSignal | null = null, authConfig = AuthConfig.WITSML_AUTHENTICATION_REQUIRED): Promise<Response> {
-    const currentCredentials = CredentialsService.getCredentials();
-    const requestInit = {
-      signal: abortSignal,
-      method: "DELETE",
-      headers: ApiClient.getCommonHeaders(currentCredentials)
-    };
-
-    return ApiClient.runHttpRequest(pathName, requestInit, authConfig);
-  }
-
-  private static runHttpRequest(pathName: string, requestInit: RequestInit, authConfig: AuthConfig) {
-    return new Promise<Response>((resolve, reject) => {
-      if (authConfig === AuthConfig.WITSML_AUTHENTICATION_REQUIRED && !("Authorization" in requestInit.headers)) {
-        reject("Not authorized");
-      }
-
-      const url = new URL(ApiClient.getBasePathName() + pathName, ApiClient.getBaseUrl());
-
-      fetch(url.toString(), requestInit)
-        .then((response) => resolve(response))
-        .catch((error) => {
-          if (error.name === "AbortError") {
-            return;
-          }
-          reject(error);
-        });
-    });
-  }
-
-  private static getBasePathName(): string {
-    const basePathName = ApiClient.getBaseUrl().pathname;
-    return basePathName !== "/" ? basePathName : "";
-  }
-
-  public static getBaseUrl(): URL {
-    let baseUrl: URL;
-    try {
-      const configuredUrl = process.env.WITSMLEXPLORER_API_URL;
-      if (configuredUrl && configuredUrl.length > 0) {
-        baseUrl = new URL(configuredUrl);
-      } else {
-        const protocol = window.location.protocol.slice(0, -1);
-        const host = window.location.hostname;
-        const port = window.location.port === "3000" ? ":5000" : "";
-        baseUrl = new URL(`${protocol}://${host}${port}`);
-      }
-    } catch (e) {
-      baseUrl = new URL("http://localhost");
+export function getBaseUrl(): URL {
+  let baseUrl: URL;
+  try {
+    const configuredUrl = process.env.WITSMLEXPLORER_API_URL;
+    if (configuredUrl && configuredUrl.length > 0) {
+      baseUrl = new URL(configuredUrl);
+    } else {
+      const protocol = window.location.protocol.slice(0, -1);
+      const host = window.location.hostname;
+      const port = window.location.port === "3000" ? ":5000" : "";
+      baseUrl = new URL(`${protocol}://${host}${port}`);
     }
-    return baseUrl;
+  } catch (e) {
+    baseUrl = new URL("http://localhost");
   }
+  return baseUrl;
 }
 
 export function truncateAbortHandler(e: Error): void {
