+++ Src/WitsmlExplorer.Frontend/services/wellboreService.tsx
@@ -1,11 +1,11 @@
-import ApiClient from "./apiClient";
 import Wellbore, { emptyWellbore } from "../models/wellbore";
-import LogObjectService from "./logObjectService";
-import TrajectoryService from "./trajectoryService";
+import { ApiClient } from "./apiClient";
 import BhaRunService from "./bhaRunService";
+import LogObjectService from "./logObjectService";
 import RigService from "./rigService";
-import TubularService from "./tubularService";
 import RiskObjectService from "./riskObjectService";
+import TrajectoryService from "./trajectoryService";
+import TubularService from "./tubularService";
 import WbGeometryObjectService from "./wbGeometryService";
 
 export default class WellboreService {
